+++ example/lib/example5_themes_playground/widgets/panels/introduction/introduction_panel.dart
@@ -3,6 +3,7 @@ import 'package:flutter/material.dart';
 import '../../../../shared/const/app.dart';
 import '../../../../shared/controllers/theme_controller.dart';
 import '../../../../shared/utils/link_text_span.dart';
+import '../../../../shared/widgets/universal/list_tile_reveal.dart';
 import '../../dialogs/show_copy_setup_code_dialog.dart';
 
 /// Show the main panel with introduction information.
@@ -38,11 +39,16 @@ class IntroductionPanel extends StatelessWidget {
     final TextStyle linkStyle =
         App.notoSansBold.copyWith(color: theme.colorScheme.primary);
 
+    final TextStyle spanSmallTextStyle = theme.textTheme.bodySmall!;
+    final TextStyle linkSmallStyle = theme.textTheme.bodySmall!.copyWith(
+        color: theme.colorScheme.primary, fontWeight: FontWeight.bold);
+
     final EdgeInsetsGeometry listTilePadding = theme.useMaterial3
         ? const EdgeInsetsDirectional.only(
             start: 16.0, end: 24.0, top: 8, bottom: 8)
         : const EdgeInsets.symmetric(horizontal: 16, vertical: 8);
 
+    // TODO(rydmike): Make one more review of the intro text.
     return Column(
       children: <Widget>[
         const SizedBox(height: 8),
@@ -71,9 +77,9 @@ class IntroductionPanel extends StatelessWidget {
                 ),
                 TextSpan(
                   style: spanTextStyle,
-                  text: ' this Playground, give it a read. It not '
-                      'only explains this application and FlexColorScheme API, '
-                      'with tutorial examples and an API guide, it also '
+                  text: ' using Themes Playground, give it a read. It not '
+                      "'only explains this application and FlexColorScheme's "
+                      'API with tutorials, examples and an API guide, it also '
                       'contains general Flutter theming guidance and advice.',
                 ),
               ],
@@ -82,13 +88,15 @@ class IntroductionPanel extends StatelessWidget {
         ),
         const ListTile(
           subtitle: Text(
-            'This application does not adjust any individual widget '
-            'properties. The theme is adjusted interactively and all Flutter '
-            "widgets in it change as the application's overall theme is "
-            'modified interactively.\n'
+            'An interesting feature of this application is that it does NOT '
+            'adjust any individual widget properties for its styling. '
+            "As you design your theme, this application's theme is also "
+            'adjusted interactively and all Flutter '
+            "widgets in it change, as the application's overall theme is "
+            'modified.\n'
             '\n'
-            'Material-3 ColorScheme is supported and can be '
-            'generated using built-in '
+            'Material-3 and its complicated ColorScheme is fully supported '
+            'and can be generated using built-in '
             'scheme colors as ColorScheme seed keys. Like when using '
             'ColorScheme.fromSeed, you can use the primary color as seed key. '
             'You can also include secondary and tertiary colors as key colors '
@@ -104,16 +112,10 @@ class IntroductionPanel extends StatelessWidget {
             'the Flutter official Material-3 demo app. You can try these apps '
             'on many different mock devices.\n'
             '\n'
-            'This app has two different modes with different layouts. '
-            'The default one is a page view, '
-            'where you see one, or in wider screens, two selectable topics '
-            'at a time. The second screen is a '
-            'large masonry grid view, where you can see many topics at once '
-            'when using very large screens. You can open and close topics to '
-            'view more or fewer of them at the same time. '
-            'This app is responsive down to phone sizes, but is primarily '
-            'intended to be used on web based desktop browsers, or as a '
-            'native Windows, Mac or Linux desktop app, where it shines.\n'
+            'This app is responsive down to larger phone sizes, but is '
+            'primarily intended to be used on web based desktop browsers, or '
+            'even better, as a native Windows, Mac or Linux desktop app, '
+            'where it shines.\n'
             '\n'
             'The first three color sets in the app are example custom color '
             'schemes. They are not built-in choices in the package. In the '
@@ -147,33 +149,31 @@ class IntroductionPanel extends StatelessWidget {
           onChanged: controller.setUseFlexColorScheme,
         ),
         SwitchListTile(
-          subtitle: const Text('The opinionated component sub-theming '
-              'is ON by default in the Playground, but OFF by default in '
+          subtitle: const Text('The component themes are '
+              'ON by default in the Playground, but OFF by default in '
               'the API. By using them you get pre-styled components that you '
-              'can adjust further.'),
+              'can adjust further to your liking.'),
           value: controller.useSubThemes && controller.useFlexColorScheme,
           onChanged:
               controller.useFlexColorScheme ? controller.setUseSubThemes : null,
         ),
         const ListTile(
           subtitle: Text(
-            'When you enable component sub-themes you get an only slightly '
-            'opinionated Material-3 default setup. If you turn OFF Material-3, '
-            'you get a Material-2 theme that is inspired by Material-3, '
-            'without using the "useMaterial3" toggle in ThemeData. All widgets '
-            'are then M2 based, you can still give them colors from the M3 '
-            'ColorScheme and also use M3 style seeded ColorScheme. ',
+            'When you enable component themes in Material-3 mode, you get a '
+            'mostly Material-3 default spec design. If you turn OFF '
+            'Material-3 mode, by default you get a Material-2 theme that is '
+            'heavily inspired by Material-3 design.',
           ),
         ),
         const ListTile(
           subtitle: Text('Like Flutter, FlexColorScheme uses different '
-              'theming defaults for properties in M2 and M3 mode. '
-              'If a value has not been set, the Playground will show you what '
-              'the default value is. When you toggle M2/M3 the value will '
-              'change and show the default '
+              'theming defaults for properties in Material-3 and Material-2 '
+              'mode. If a value has not been set, the Playground will show '
+              'you what the default value is. When you toggle the mode, '
+              'the value will change and show the default '
               'for the used mode. The Playground also shows what the default '
-              'values are in Flutter SDK in both M2 and M3 mode, when '
-              'FlexColorScheme or its sub-themes are not being used.'),
+              'values are in Flutter SDK in both Material-3 and Material-2 '
+              'mode, when FlexColorScheme or its sub-themes are not being used.'),
         ),
         ListTile(
           subtitle: const Text(
@@ -196,80 +196,62 @@ class IntroductionPanel extends StatelessWidget {
               : null,
         ),
         const Divider(),
-        const ListTile(
-          title: Text('Material-3 will be made default'),
-          subtitle: Text('Flutter SDK will be switching to using M3 by default '
-              'later during 2023, even when not setting the '
-              '"ThemeData.useMaterial3" flag to true. After that point, '
-              'if you intend to use Material-2, you have to '
-              'explicitly opt-out of using Material-3 by setting the '
-              'flag to false. In anticipation of this change, the '
-              'Themes Playground code gen now always generates the code '
-              'config for selected Material M2/M3 mode to always make it clear '
-              'what the intent is.'),
-        ),
         SwitchListTile(
-          title: const Text('Use Material-3 design'),
+          title: const Text('Material-3 design'),
+          subtitle: const Text(
+              'Flutter SDK and FlexColorScheme use Material-3 design '
+              'styled components by default.'),
           value: controller.useMaterial3,
           onChanged: controller.setUseMaterial3,
         ),
         const ListTile(
           title: Text('Material-2 will be deprecated'),
           subtitle: Text(
-              'Another coming big change is the deprecation of Material-2 '
-              'and eventual removal of it from the Flutter SDK '
-              'code-base. If you use M2, you might want to consider migrating '
-              'to M3 soonish. Incidentally, with FlexColorScheme and '
-              'Themes Playground, you can  make a theme that is '
-              'based on Material-3, but looks like Material-2. There '
-              'is even a premade design to demonstrate this in the '
-              'Playground. As such the M2 design is probably not something '
+              'Legacy Material-2 support is still supported in Flutter 3.24 '
+              'and FCS V8, but it will eventually be deprecated and '
+              'removed from the Flutter SDK code-base. If you still use '
+              'Material-2, you should migrate to run your Flutter app '
+              'in Material-3 mode. '
+              'If you prefer the Material-2 style, using FlexColorScheme '
+              'and the Themes Playground, you can make a theme that is '
+              'based on Material-3, but looks very close to Material-2. There '
+              'is a premade design to demonstrate this in the '
+              'Playground. The Material-2 design is probably not something '
               'to strive for anymore, but this demonstrates the flexibility '
-              'of Material-3 design in Flutter, and the powerful theming '
-              'features in the Themes Playground.'),
+              "of Material-3 design's theming in Flutter, and the powerful "
+              'theming features available in the Themes Playground.'),
         ),
         const ListTile(
-          title: Text('Flutter SDK Material 3 Issues'),
-          subtitle: Text('In Flutter 3.7 and later there are only minor issues '
-              'left when using Material-3. The issues have a very low impact '
-              'on apps using Material-3. '
-              'FlexColorScheme is able to work around and fix some of '
-              'the issues. In Flutter 3.3 and earlier, there are significant '
-              'issues related to enabling and using Material-3. Do not use '
-              'Material-3 with Flutter 3.3 or earlier. For the best '
-              'experience when using Material-3, prefer using Flutter 3.10 '
-              'or later.'),
+          title: Text('Known issues'),
+          subtitle: Text('There are many known issues and theming '
+              'limitations in Flutter. In the Playground, you will find '
+              'information about the most common ones and limitations '
+              "presented next to each component's theming settings. "
+              'The issues are presented in an info expand like the one shown '
+              'below. You can view it by clicking on the small info icon.'),
         ),
-        Padding(
-          padding: listTilePadding,
-          child: RichText(
+        ListTileReveal(
+          dense: true,
+          title: const Text('Known issues'),
+          subtitleReveal: RichText(
             text: TextSpan(
               children: <TextSpan>[
                 TextSpan(
-                  style: spanTextStyle,
-                  text: 'You can read more about M3 and Flutter theming '
-                      'related issues in the FlexColorScheme docs, where '
-                      'Flutter theming issues are ',
+                  style: spanSmallTextStyle,
+                  text: 'Info about known theming issues in Flutter are '
+                      'presented in component theming settings in this style. '
+                      'You can also read more about Flutter '
+                      'theming related issues in the FlexColorScheme docs. '
+                      'Key Flutter theming issues are ',
                 ),
                 LinkTextSpan(
-                  style: linkStyle,
+                  style: linkSmallStyle,
                   uri: _fcsFlutterIssues,
-                  text: 'tracked',
+                  text: 'tracked here',
                 ),
                 TextSpan(
-                  style: spanTextStyle,
-                  text: '. Due to the large number of issues and also '
-                      'incomplete state of Material-3 implementation in '
-                      'Flutter 3.3 and earlier, do not use '
-                      'M3 with Flutter 3.3 or earlier versions. '
-                      'However, in Flutter 3.7 or later versions, M3 is '
-                      'ready and can be used in production. Some M3 '
-                      'fidelity and specification compliance issues still '
-                      'exist, even in 3.10 and later. Even if there are '
-                      'quite a large number of issues, they are typically '
-                      'minor and will be fixed in later releases. Many are '
-                      'fixed in master channel and some already in beta '
-                      'channel, waiting for next stable release after 3.10.\n',
+                  style: spanSmallTextStyle,
+                  text: '.\n',
                 ),
               ],
             ),